requirements.txt
Django==4.2
python-dotenv
requests
pillow
pixelmatch
playwright
django-crispy-forms
crispy-bootstrap5
